+++ CHANGELOG.md
@@ -7,6 +7,8 @@
 * Add `MonadRWS` instance.
 * Add `MonadAccum` and `MonadSelect` instances, whenever mtl-2.3 or newer is used.
 * Add `MonadThrow` instance.
+* Fix a bug in `Applicative` instance. Implementation for `*>` did not short circuit
+  mondic computation.
 
 ## 0.1.0.0